pet_alert/requirements.txt
Django==4.2.1
django-filter==23.2
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
django-registration==3.3
gunicorn==20.1.0
phonenumberslite==8.13.11
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==0.21.1
sorl-thumbnail==12.9.0
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2